feat: add `/conductor` submodule
