+++ workspace/aubade/package.json
@@ -35,6 +35,10 @@
 			"types": "./index.d.ts",
 			"default": "./src/compass/index.js"
 		},
+		"./conductor": {
+			"types": "./index.d.ts",
+			"default": "./src/conductor/index.js"
+		},
 		"./transform": {
 			"types": "./index.d.ts",
 			"default": "./src/transform/index.js"
