+++ workspace/aubade/src/conductor/index.test.ts
@@ -1,11 +1,11 @@
-import { traverse } from './index.js';
+import { assemble } from './index.js';
 
 declare function expect<T>(v: T): void;
 
 async (/* traverse */) => {
-	expect<Record<string, any>>(await traverse('.'));
+	expect<Record<string, any>>(await assemble('.'));
 
-	const [item] = await traverse('.', ({ breadcrumb: [, slug] }) => {
+	const [item] = await assemble('.', ({ breadcrumb: [, slug] }) => {
 		if (!slug.endsWith('.md')) return;
 		return async ({ buffer, marker, parse }) => {
 			const { body, frontmatter } = parse(buffer.toString('utf-8'));
